requirements.txt
joblib==1.5.1
mlflow==3.1.0
mlflow_skinny==3.1.0
nltk==3.9.1
numpy<2.3.0
pandas<2.3.0
scikit_learn==1.7.0
python-dotenv==1.0.0
psycopg2-binary
requests
schedule
sqlalchemy